Reformat /test/testit/*.R
In assert() forms, yank description to the same line; use a single expression wrapped by {} as the second argument. Within this expression, wrap in ()s any sub-expressions which would evaluate to a logical value and be used for the assertion, including those who are returned by vapply() or Map(), as well as `TRUE`s.